Rename trainer subfiles and add additional init files
