+++ python/unitytrainers/__init__.py
@@ -1,7 +1,7 @@
 from .buffer import *
 from .models import *
 from .trainer_controller import *
-from .bc.bc_models import *
-from .bc.bc_trainer import *
-from .ppo.ppo_models import *
-from .ppo.ppo_trainer import *
+from .bc.models import *
+from .bc.trainer import *
+from .ppo.models import *
+from .ppo.trainer import *
