+++ python/unitytrainers/bc/__init__.py
@@ -1,0 +1,2 @@
+from .models import *
+from .trainer import *
